Move createSampleGeneratorContext to test-utils
