+++ generators/swift/model/src/object/__test__/object-generator.test.ts
@@ -1,9 +1,9 @@
+import { resolve } from "node:path";
 import { SwiftFile } from "@fern-api/swift-base";
 import { swift } from "@fern-api/swift-codegen";
-
 import { ModelGeneratorContext } from "../../ModelGeneratorContext";
 import { ObjectGenerator } from "../../object";
-import { createSampleGeneratorContext } from "./util/createSampleGeneratorContext";
+import { createSampleGeneratorContext } from "../../test-utils/createSampleGeneratorContext";
 
 function getObjectTypeDeclarationOrThrow(context: ModelGeneratorContext, name: string) {
     for (const declaration of Object.values(context.ir.types)) {
@@ -22,9 +22,13 @@ function getObjectTypeDeclarationOrThrow(context: ModelGeneratorContext, name: s
     throw new Error(`Type declaration for ${name} not found`);
 }
 
+function pathToDefinition(testDefinitionName: string) {
+    return resolve(__dirname, "./test-definitions", testDefinitionName);
+}
+
 describe("ObjectGenerator", () => {
     it("correctly generates nested enums for duplicate string literal values", async () => {
-        const context = await createSampleGeneratorContext("duplicate-string-literals");
+        const context = await createSampleGeneratorContext(pathToDefinition("duplicate-string-literals"));
         const objectName = "ObjectWithDuplicateStringLiterals";
         const declaration = getObjectTypeDeclarationOrThrow(context, objectName);
         const generator = new ObjectGenerator({
@@ -38,7 +42,7 @@ describe("ObjectGenerator", () => {
     });
 
     it(`ensures that the special 'CodingKeys' enum does not collide with other string literal enums`, async () => {
-        const context = await createSampleGeneratorContext("coding-keys-literal");
+        const context = await createSampleGeneratorContext(pathToDefinition("coding-keys-literal"));
         const objectName = "ObjectWithCodingKeysLiteral";
         const declaration = getObjectTypeDeclarationOrThrow(context, objectName);
         const generator = new ObjectGenerator({
@@ -52,7 +56,7 @@ describe("ObjectGenerator", () => {
     });
 
     it(`correctly generates literals in container types`, async () => {
-        const context = await createSampleGeneratorContext("literals-in-container-types");
+        const context = await createSampleGeneratorContext(pathToDefinition("literals-in-container-types"));
         const objectName = "ObjectWithLiteralsInContainerTypes";
         const declaration = getObjectTypeDeclarationOrThrow(context, objectName);
         const generator = new ObjectGenerator({
@@ -66,7 +70,7 @@ describe("ObjectGenerator", () => {
     });
 
     it(`correctly handles name conflicts between nested types and schema types`, async () => {
-        const context = await createSampleGeneratorContext("nested-type-collision-with-schema-type");
+        const context = await createSampleGeneratorContext(pathToDefinition("nested-type-collision-with-schema-type"));
 
         const fileComponents: swift.FileComponent[] = [];
