+++ generators/swift/model/src/test-utils/createSampleGeneratorContext.ts
@@ -1,14 +1,12 @@
-import { resolve } from "node:path";
-
 import { FernGeneratorExec, GeneratorNotificationService } from "@fern-api/base-generator";
 import { AbsoluteFilePath } from "@fern-api/fs-utils";
 import { createSampleIr } from "@fern-api/test-utils";
 import { IntermediateRepresentation } from "@fern-fern/ir-sdk/api";
-import { ModelCustomConfigSchema } from "../../../ModelCustomConfig";
-import { ModelGeneratorContext } from "../../../ModelGeneratorContext";
+import { ModelCustomConfigSchema } from "../ModelCustomConfig";
+import { ModelGeneratorContext } from "../ModelGeneratorContext";
 
-export async function createSampleGeneratorContext(testDefinitionName: string): Promise<ModelGeneratorContext> {
-    const absolutePathToWorkspace = AbsoluteFilePath.of(resolve(__dirname, "../test-definitions", testDefinitionName));
+export async function createSampleGeneratorContext(pathToDefinition: string): Promise<ModelGeneratorContext> {
+    const absolutePathToWorkspace = AbsoluteFilePath.of(pathToDefinition);
     const ir = await createSampleIr(absolutePathToWorkspace);
     const generatorConfig = createSampleGeneratorConfig();
     const customConfig: ModelCustomConfigSchema = {};
